spot support

+++ src/main/java/io/contek/invoker/deribit/api/common/_Position.java
@@ -1,32 +1,33 @@
 package io.contek.invoker.deribit.api.common;
 
 import javax.annotation.concurrent.NotThreadSafe;
+import java.math.BigDecimal;
 
 @NotThreadSafe
 public class _Position {
 
-  public Double average_price;
-  public Double average_price_usd;
-  public Double delta;
+  public BigDecimal average_price;
+  public BigDecimal average_price_usd;
+  public BigDecimal delta;
   public String direction;
-  public Double estimated_liquidation_price;
-  public Double floating_profit_loss;
-  public Double floating_profit_loss_usd;
-  public Double gamma;
-  public Double index_price;
-  public Double initial_margin;
+  public BigDecimal estimated_liquidation_price;
+  public BigDecimal floating_profit_loss;
+  public BigDecimal floating_profit_loss_usd;
+  public BigDecimal gamma;
+  public BigDecimal index_price;
+  public BigDecimal initial_margin;
   public String instrument_name;
   public String kind;
   public Integer leverage;
-  public Double maintenance_margin;
-  public Double mark_price;
-  public Double open_orders_margin;
-  public Double realized_funding;
-  public Double realized_profit_loss;
-  public Double settlement_price;
-  public Integer size;
+  public BigDecimal maintenance_margin;
+  public BigDecimal mark_price;
+  public BigDecimal open_orders_margin;
+  public BigDecimal realized_funding;
+  public BigDecimal realized_profit_loss;
+  public BigDecimal settlement_price;
+  public BigDecimal size;
   public String size_currency;
-  public Double theta;
-  public Double total_profit_loss;
-  public Double vega;
+  public BigDecimal theta;
+  public BigDecimal total_profit_loss;
+  public BigDecimal vega;
 }
